Refactor the CLI to address memory issues
The old CLI was designed with 1 location in mind.
Now that 33 locations are supported, trying to hold all data for all locations in memory simultaneously does not scale.
Instead, this commit focuses on 2 things:
* Do 1 operation (per scraper) per execution.
* Do not aggregate data across scrapers.
This means all locations are not plottable at once, but that is ok since the labels already overlap each other to the point of illegibilty.